+++ src/string/index.js
@@ -1,4 +1,3 @@
 export * from './text';
 export * from './url';
 export * from './naming';
-